Allow SQLAlchemy w/ Django tests + CTE unit tests
