+++ specifyweb/settings/__init__.py
@@ -56,12 +56,17 @@ DATABASES = {
     },
  }
 
-SA_DATABASE_URL = 'mysql://%s:%s@%s:%s/%s?charset=utf8' % (
+def get_sa_db_url(db_name):
+    return 'mysql://%s:%s@%s:%s/%s?charset=utf8' % (
         MASTER_NAME,
         MASTER_PASSWORD,
         DATABASE_HOST,
         DATABASE_PORT or 3306,
-        DATABASE_NAME)
+        db_name)
+
+SA_DATABASE_URL = get_sa_db_url(DATABASE_NAME)
+
+SA_TEST_DB_URL = get_sa_db_url(f'test_{DATABASE_NAME}')
 
 # Prevent MySQL connection timeouts
 SA_POOL_RECYCLE = 3600
