+++ specifyweb/workbench/upload/tests/testuploading.py
@@ -9,8 +9,10 @@ from jsonschema import validate  # type: ignore
 
 from specifyweb.specify import auditcodes
 from specifyweb.specify.auditlog import auditlog
+from specifyweb.specify.test_trees import TestTree
 from specifyweb.specify.tree_extras import validate_tree_numbering
-from .base import UploadTestsBase, get_table
+from specifyweb.specify.test_trees import get_table
+from .base import UploadTestsBase
 from ..parsing import filter_and_upload
 from ..tomany import ToManyRecord
 from ..treerecord import TreeRecord, BoundTreeRecord, \
@@ -23,100 +25,11 @@ from ..upload_table import UploadTable, ScopedUploadTable, \
     _to_many_filters_and_excludes, BoundUploadTable
 from ..uploadable import Exclude, Auditor
 
+class UploadTreeSetup(TestTree, UploadTestsBase): pass
 
-class TreeMatchingTests(UploadTestsBase):
+class TreeMatchingTests(UploadTreeSetup):
     def setUp(self) -> None:
         super().setUp()
-
-        self.earth = get_table('Geography').objects.create(
-            name="Earth",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="Planet"),
-            definition=self.geographytreedef,
-        )
-
-        self.na = get_table('Geography').objects.create(
-            name="North America",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="Continent"),
-            definition=self.geographytreedef,
-            parent=self.earth,
-        )
-
-        self.usa = get_table('Geography').objects.create(
-            name="USA",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="Country"),
-            definition=self.geographytreedef,
-            parent=self.na,
-        )
-
-        self.kansas = get_table('Geography').objects.create(
-            name="Kansas",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="State"),
-            definition=self.geographytreedef,
-            parent=self.usa,
-        )
-
-        self.mo = get_table('Geography').objects.create(
-            name="Missouri",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="State"),
-            definition=self.geographytreedef,
-            parent=self.usa,
-        )
-
-        self.ohio = get_table('Geography').objects.create(
-            name="Ohio",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="State"),
-            definition=self.geographytreedef,
-            parent=self.usa,
-        )
-
-        self.ill = get_table('Geography').objects.create(
-            name="Illinois",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="State"),
-            definition=self.geographytreedef,
-            parent=self.usa,
-        )
-
-        self.doug = get_table('Geography').objects.create(
-            name="Douglas",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="County"),
-            definition=self.geographytreedef,
-            parent=self.kansas,
-        )
-
-        self.greene = get_table('Geography').objects.create(
-            name="Greene",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="County"),
-            definition=self.geographytreedef,
-            parent=self.mo,
-        )
-
-        self.greeneoh = get_table('Geography').objects.create(
-            name="Greene",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="County"),
-            definition=self.geographytreedef,
-            parent=self.ohio,
-        )
-
-        self.sangomon = get_table('Geography').objects.create(
-            name="Sangamon",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="County"),
-            definition=self.geographytreedef,
-            parent=self.ill,
-        )
-
-        self.springmo = get_table('Geography').objects.create(
-            name="Springfield",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="City"),
-            definition=self.geographytreedef,
-            parent=self.greene,
-        )
-
-        self.springill = get_table('Geography').objects.create(
-            name="Springfield",
-            definitionitem=get_table('Geographytreedefitem').objects.get(name="City"),
-            definition=self.geographytreedef,
-            parent=self.sangomon,
-        )
 
     def test_enforced_state(self) -> None:
         state = get_table('Geographytreedefitem').objects.get(name="State")
